Addition of Leave command
If there's a join command, there should be a leave command too, no?
